# docs/openapi.yaml
openapi: 3.1.0
info:
  title: energeX API
  version: 1.0.0
  description: A microservice API using Lumen (PHP) and Node.js (TypeScript) that integrates with Redis for caching, a MySQL database, and a simple frontend (React.js or Vue.js) to consume the API.
servers:
- url: http://localhost:8080
  description: Lumen (php) - serves /api/*
- url: http://localhost:4000
  description: cache-node – serves /cache/*
...
paths:
...
  /cache/posts:
    get:
      tags: [Cache]
      summary: Fetch all posts (from cache-node)
      responses:
        "200":
          description: A list of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      format: int64
                    title:
                      type: string
                    content:
                      type: string
                    user_id:
                      type: integer
                    created_at:
                      type: string
                      format: date-time
        "500":
          description: Server error
  /cache/posts/{id}:
    get:
      tags: [Cache]
      summary: Fetch a single post(from cache-node)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: A single post
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                  title:
                    type: string
                  content:
                    type: string
                  user_id:
                    type: integer
                  created_at:
                    type: string
                    format: date-time
        "404":
          description: Post not found
        "500":
          description: Server error          
components:
  securitySchemes: 
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...